+++ exonum/src/api/public/blockchain_explorer.rs
@@ -23,7 +23,6 @@ use api::{Api, ApiError};
 use blockchain::{Block, Blockchain, TxLocation, Schema, TransactionErrorType, TransactionResult};
 use crypto::Hash;
 use helpers::Height;
-use node::state::TxPool;
 use messages::Precommit;
 use storage::ListProof;
 
@@ -101,13 +100,12 @@ pub enum TransactionInfo {
 #[derive(Clone, Debug)]
 pub struct ExplorerApi {
     blockchain: Blockchain,
-    pool: TxPool,
 }
 
 impl ExplorerApi {
     /// Creates a new `ExplorerApi` instance.
-    pub fn new(pool: TxPool, blockchain: Blockchain) -> Self {
-        ExplorerApi { pool, blockchain }
+    pub fn new(blockchain: Blockchain) -> Self {
+        ExplorerApi { blockchain }
     }
 
     fn explorer(&self) -> BlockchainExplorer {
@@ -134,10 +132,21 @@ impl ExplorerApi {
     }
 
     fn transaction_info(&self, hash: &Hash) -> Result<TransactionInfo, ApiError> {
-        if let Some(tx) = self.pool.read().expect("Unable to read pool").get(hash) {
-            Ok(TransactionInfo::InPool {
-                content: tx.serialize_field().map_err(ApiError::InternalError)?,
-            })
+        let snapshot = self.blockchain.snapshot();
+        let schema = Schema::new(&snapshot);
+        if schema.transactions_pool().contains(hash) {
+            let raw_tx = schema.transactions().get(hash).expect(
+                "Expected tx in database",
+            );
+
+            let box_transaction = self.blockchain.tx_from_raw(raw_tx.clone()).ok_or_else(|| {
+                ApiError::InternalError(format!("Service not found for tx: {:?}", raw_tx).into())
+            })?;
+
+            let content = box_transaction.serialize_field().map_err(
+                ApiError::InternalError,
+            )?;
+            Ok(TransactionInfo::InPool { content: content })
         } else if let Some(tx_info) = self.explorer().tx_info(hash)? {
             Ok(TransactionInfo::Committed(tx_info))
         } else {