+++ exonum/src/node/consensus.rs
@@ -117,28 +117,24 @@ impl NodeHandler {
             return;
         }
 
+        trace!("Handle propose");
+
         let snapshot = self.blockchain.snapshot();
-        // Check that transactions are not committed yet
-        for hash in msg.transactions() {
-            if Schema::new(&snapshot).transactions().contains(hash) {
-                error!(
-                    "Received propose with already committed transaction, msg={:?}",
-                    msg
-                );
+        let schema = Schema::new(snapshot);
+        //TODO: remove this match after errors refactor.
+        let has_unknown_txs = match self.state.add_propose(
+            msg,
+            schema.transactions(),
+            schema.transactions_pool(),
+        ) {
+            Ok(state) => state.has_unknown_txs(),
+            Err(err) => {
+                warn!("{}, msg={:?}", err, msg);
                 return;
             }
-        }
-
-        if self.state.propose(&msg.hash()).is_some() {
-            return;
-        }
-
-        trace!("Handle propose");
-        // Add propose
-        let (hash, has_unknown_txs) = match self.state.add_propose(msg) {
-            Some(state) => (state.hash(), state.has_unknown_txs()),
-            None => return,
         };
+
+        let hash = msg.hash();
 
         // Remove request info
         let known_nodes = self.remove_request(&RequestData::Propose(hash));
@@ -210,34 +206,42 @@ impl NodeHandler {
         }
 
         if self.state.block(&block_hash).is_none() {
-            let snapshot = self.blockchain.snapshot();
-            let schema = Schema::new(&snapshot);
+
+
             // Verify transactions
             let mut tx_hashes = Vec::new();
-            for raw in msg.transactions() {
-                if let Some(tx) = self.blockchain.tx_from_raw(raw) {
-                    let hash = tx.hash();
-                    if schema.transactions().contains(&hash) {
-                        error!(
-                            "Received block with already committed transaction, block={:?}",
-                            msg
-                        );
-                        return;
-                    }
-                    profiler_span!("tx.verify()", {
-                        if !tx.verify() {
-                            error!("Incorrect transaction in block detected, block={:?}", msg);
+            let mut fork = self.blockchain.fork();
+            {
+                let mut schema = Schema::new(&mut fork);
+
+                for raw in msg.transactions() {
+                    if let Some(tx) = self.blockchain.tx_from_raw(raw) {
+                        let hash = tx.hash();
+                        if schema.transactions().get(&hash).is_some() {
+                            error!(
+                                "Received block with already committed transaction, block={:?}",
+                                msg
+                            );
                             return;
                         }
-                    });
-                    self.state.add_transaction(hash, tx, true);
-                    tx_hashes.push(hash);
-                } else {
-                    error!("Unknown transaction in block detected, block={:?}", msg);
-                    return;
+                        profiler_span!("tx.verify()", {
+                            if !tx.verify() {
+                                error!("Incorrect transaction in block detected, block={:?}", msg);
+                                return;
+                            }
+                        });
+                        schema.transactions_pool_mut().insert(hash);
+                        schema.transactions_mut().put(&hash, tx.raw().clone());
+                        tx_hashes.push(hash);
+                    } else {
+                        error!("Unknown transaction in block detected, block={:?}", msg);
+                        return;
+                    }
                 }
             }
-
+            self.blockchain.merge(fork.into_patch()).expect(
+                "Unable to save transaction to persistent pool.",
+            );
             let (block_hash, patch) =
                 self.create_block(block.proposer_id(), block.height(), tx_hashes.as_slice());
             // Verify block_hash
@@ -259,6 +263,7 @@ impl NodeHandler {
         }
         self.commit(block_hash, msg.precommits().iter(), None);
         self.request_next_block();
+
     }
 
     /// Executes and commits block. This function is called when node has full propose information.
@@ -487,12 +492,11 @@ impl NodeHandler {
             );
             (block_state.txs().len(), block_state.proposer_id())
         };
+        let snapshot = self.blockchain.snapshot();
+        let schema = Schema::new(&snapshot);
+        let pool = schema.transactions_pool();
+        let mempool_size = pool.iter().count();
 
-        let mempool_size = self.state
-            .transactions()
-            .read()
-            .expect("Expected read lock")
-            .len();
         metric!("node.mempool", mempool_size);
 
         let height = self.state.height();
@@ -536,7 +540,7 @@ impl NodeHandler {
         let hash = msg.hash();
         let tx = {
             let service_id = msg.service_id();
-            if let Some(tx) = self.blockchain.tx_from_raw(msg) {
+            if let Some(tx) = self.blockchain.tx_from_raw(msg.clone()) {
                 tx
             } else {
                 error!(
@@ -548,15 +552,6 @@ impl NodeHandler {
         };
 
         profiler_span!("Make sure that it is new transaction", {
-            if self.state
-                .transactions()
-                .read()
-                .expect("Expected read lock")
-                .contains_key(&hash)
-            {
-                return;
-            }
-
             let snapshot = self.blockchain.snapshot();
             if Schema::new(&snapshot).transactions().contains(&hash) {
                 return;
@@ -568,8 +563,17 @@ impl NodeHandler {
                 return;
             }
         });
+        let mut fork = self.blockchain.fork();
+        {
+            let mut schema = Schema::new(&mut fork);
+            schema.transactions_pool_mut().insert(hash);
+            schema.transactions_mut().put(&hash, msg.raw().clone());
+        }
+        self.blockchain.merge(fork.into_patch()).expect(
+            "Unable to save transaction to persistent pool.",
+        );
 
-        let full_proposes = self.state.add_transaction(hash, tx, false);
+        let full_proposes = self.state.transaction_validated(hash);
         // Go to has full propose if we get last transaction
         for (hash, round) in full_proposes {
             self.remove_request(&RequestData::Transactions(hash));
@@ -582,27 +586,20 @@ impl NodeHandler {
     pub fn handle_incoming_tx(&mut self, msg: Box<Transaction>) {
         trace!("Handle incoming transaction");
         let hash = msg.hash();
-
-        // Make sure that it is new transaction
-        if self.state
-            .transactions()
-            .read()
-            .expect("Expected read lock")
-            .contains_key(&hash)
+        let mut fork = self.blockchain.fork();
         {
-            return;
+            let mut schema = Schema::new(&mut fork);
+            schema.transactions_pool_mut().insert(hash);
+            schema.transactions_mut().put(&hash, msg.raw().clone());
         }
-
-        let snapshot = self.blockchain.snapshot();
-        if Schema::new(&snapshot).transactions().contains(&hash) {
-            return;
-        }
-
+        self.blockchain.merge(fork.into_patch()).expect(
+            "Unable to save transaction to persistent pool.",
+        );
         // Broadcast transaction to validators
         trace!("Broadcast transactions: {:?}", msg.raw());
         self.broadcast(msg.raw());
 
-        let full_proposes = self.state.add_transaction(hash, msg, false);
+        let full_proposes = self.state.transaction_validated(hash);
         // Go to has full propose if we get last transaction
         for (hash, round) in full_proposes {
             self.remove_request(&RequestData::Transactions(hash));
@@ -684,24 +681,17 @@ impl NodeHandler {
             if self.state.have_prevote(round) {
                 return;
             }
-            let pool_len = self.state
-                .transactions()
-                .read()
-                .expect("Expected read lock")
-                .len();
+            let snapshot = self.blockchain.snapshot();
+            let schema = Schema::new(&snapshot);
+            let pool = schema.transactions_pool();
+            let pool_len = pool.iter().count();
 
             info!("LEADER: pool = {}", pool_len);
 
             let round = self.state.round();
             let max_count = ::std::cmp::min(self.txs_block_limit() as usize, pool_len);
-            let txs: Vec<Hash> = self.state
-                .transactions()
-                .read()
-                .expect("Expected read lock")
-                .keys()
-                .take(max_count)
-                .cloned()
-                .collect();
+
+            let txs: Vec<Hash> = pool.iter().take(max_count).collect();
             let propose = Propose::new(
                 validator_id,
                 self.state.height(),
@@ -796,14 +786,7 @@ impl NodeHandler {
         height: Height,
         tx_hashes: &[Hash],
     ) -> (Hash, Patch) {
-        self.blockchain.create_patch(
-            proposer_id,
-            height,
-            tx_hashes,
-            &self.state.transactions().read().expect(
-                "Expected read lock",
-            ),
-        )
+        self.blockchain.create_patch(proposer_id, height, tx_hashes)
     }
 
     /// Calls `create_block` with transactions from the corresponding `Propose` and returns the