+++ sandbox/src/sandbox.rs
@@ -20,7 +20,7 @@ use std::sync::{Arc, Mutex};
 use std::cell::{Ref, RefCell, RefMut};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-use std::collections::{BTreeMap, BinaryHeap, HashMap, HashSet, VecDeque};
+use std::collections::{BTreeSet, BinaryHeap, HashMap, HashSet, VecDeque};
 use std::iter::FromIterator;
 
 use futures::{self, Async, Future, Stream};
@@ -465,28 +465,51 @@ impl Sandbox {
     where
         I: IntoIterator<Item = &'a RawTransaction>,
     {
-        let blockchain = &self.blockchain_ref();
-        let (hashes, tx_pool) = {
-            let mut pool = BTreeMap::new();
+        let height = self.current_height();
+        let mut blockchain = self.blockchain_mut();
+        let (hashes, recover, patch) = {
             let mut hashes = Vec::new();
-            for raw in txs {
-                let tx = blockchain.tx_from_raw(raw.clone()).unwrap();
-                let hash = tx.hash();
-                hashes.push(hash);
-                pool.insert(hash, tx);
+            let mut recover = BTreeSet::new();
+            let mut fork = blockchain.fork();
+            {
+                let mut schema = Schema::new(&mut fork);
+                for raw in txs {
+                    let hash = raw.hash();
+                    hashes.push(hash);
+                    if schema.transactions().get(&hash).is_none() {
+                        recover.insert(hash);
+                        schema.transactions_mut().put(&hash, raw.clone());
+                        schema.transactions_pool_mut().insert(hash);
+                    }
+                }
             }
-            (hashes, pool)
+
+            (hashes, recover, fork.into_patch())
         };
+        blockchain.merge(patch).unwrap();
 
         let fork = {
             let mut fork = blockchain.fork();
-            let (_, patch) =
-                blockchain.create_patch(ValidatorId(0), self.current_height(), &hashes, &tx_pool);
+            let (_, patch) = blockchain.create_patch(ValidatorId(0), height, &hashes);
             fork.merge(patch);
             fork
         };
+        let patch = {
+            let mut fork = blockchain.fork();
+            {
+                let mut schema = Schema::new(&mut fork);
+                for hash in recover.into_iter() {
+                    schema.transactions_mut().remove(&hash);
+                    schema.transactions_pool_mut().remove(&hash);
+                }
+            }
+            fork.into_patch()
+        };
+
+        blockchain.merge(patch).unwrap();
         *Schema::new(&fork).last_block().state_hash()
     }
+
 
     pub fn get_proof_to_service_table(&self, service_id: u16, table_idx: usize) -> MapProof<Hash> {
         let snapshot = self.blockchain_ref().snapshot();
@@ -528,11 +551,10 @@ impl Sandbox {
     }
 
     pub fn transactions_hashes(&self) -> Vec<Hash> {
-        let node_state = self.node_state();
-        let read_lock = node_state.transactions().read().expect(
-            "Expected read lock",
-        );
-        read_lock.keys().cloned().collect()
+        let schema = Schema::new(self.blockchain_ref().snapshot());
+        let idx = schema.transactions_pool();
+        let vec = idx.iter().collect();
+        vec
     }
 
     pub fn current_round(&self) -> Round {